Ingester work in progress
